requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
dateparser==1.2.0
Django==4.2.13
django-jazzmin==3.0.0
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.5
multidict==6.1.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pytesseract==0.3.13
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
wasmtime==25.0.0
websockets==13.0.1
whitenoise==6.7.0
yarl==1.11.1
